# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "contact": {}
  },
  "paths": {
    "/api/v1/login/auth_code_login": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "登录/注册"
        ],
        "summary": "验证码登录",
        "parameters": [
          {
            "type": "string",
            "description": "邮箱",
            "name": "email",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "验证码",
            "name": "authCode",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/v1/login/get_UserInfo": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "用户"
        ],
        "summary": "获取用户权限",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/v1/login/get_user_info": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "用户"
        ],
        "summary": "获取用户信息",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/v1/login/register": {
      "post": {
        "tags": [
          "登录/注册"
        ],
        "summary": "注册",
        "parameters": [
          {
            "type": "string",
            "description": "email",
            "name": "email",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "password",
            "name": "password",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/v1/login/send_email": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "登录/注册"
        ],
        "summary": "发送邮件",
        "parameters": [
          {
            "type": "string",
            "description": "邮箱",
            "name": "email",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "验证码",
            "name": "authCode",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/types.RegisterResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/types.ErrorResponse"
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "登录/注册"
        ],
        "summary": "登录",
        "parameters": [
          {
            "type": "string",
            "description": "邮箱",
            "name": "email",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "description": "密码",
            "name": "password",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "types.ErrorResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "types.RegisterResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer"
        },
        "message": {
          "type": "string"
        },
        "token": {
          "type": "string"
        },
        "userInfo": {}
      }
    }
  }
}